Fix type casting in log message for socket error

+++ test/windows/NetworkTests.cpp
@@ -2279,7 +2279,7 @@ class NetworkTests
                     int localAddrLength = sizeof(localAddr);
                     if (getsockname(sock.get(), reinterpret_cast<SOCKADDR*>(&localAddr), &localAddrLength) == SOCKET_ERROR)
                     {
-                        LogInfo("connect() to port %u failed with %d; getsockname() failed with %d", assignedPort, connectError, WSAGetLastError());
+LogInfo("connect() to port %u failed with %d; getsockname() failed with %d", static_cast<unsigned int>(assignedPort), connectError, WSAGetLastError());
                     }
                     else
                     {
